Multivariate Regression Example
Regression goes through as many data points as possible. MAE, MSE, and RMSE, evaluate how well a model is done. MAE shows how many total errors it has. MSE takes more of the average error, which means all background data and noise has been eliminated. RMSE puts the error in the same index(unit) as Y. When we look at this, it is in the same unit as Y.